requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.12
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
idna==3.10
pydantic==2.11.2
pydantic_core==2.33.1
python-dotenv==1.1.0
RapidFuzz==3.13.0
requests==2.32.3
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
